finishing up

+++ code/modules/heavy_vehicle/mech_interaction.dm
@@ -350,9 +350,15 @@
 				if(length(pilots))
 					to_chat(user, SPAN_WARNING("You can't apply this upgrade while \the [src] has occupants!"))
 					return
+				if(!maintenance_protocols)
+					to_chat(user, SPAN_WARNING("You are unable to apply this upgrade while \the [src]'s maintenance protocols are not active."))
+					return
 				user.visible_message(SPAN_NOTICE("\The [user] begins installing \the [thing] into \the [src]..."), SPAN_NOTICE("You begin installing the [thing] into \the [src]..."))
 				if(do_after(user, 30, TRUE, src))
+					var/obj/item/remote_mecha/RM = thing
 					user.visible_message(SPAN_NOTICE("\The [user] installs \the [thing] into \the [src]."), SPAN_NOTICE("You install the [thing] into \the [src]."))
+					remote_network = RM.mech_remote_network
+					does_hardpoint_lock = RM.hardpoint_lock
 					become_remote()
 					qdel(thing)
 			else if(thing.ismultitool())
@@ -373,12 +379,17 @@
 				return
 
 			else if(thing.iswrench())
-				if(!maintenance_protocols)
-					to_chat(user, "<span class='warning'>The securing bolts are not visible while maintenance protocols are disabled.</span>")
+				if(length(pilots))
+					to_chat(user, SPAN_WARNING("You can't disassemble \the [src] while it has a pilot!"))
 					return
-				to_chat(user, "<span class='notice'>You dismantle \the [src].</span>")
-				dismantle()
-				return
+				if(!maintenance_protocols)
+					to_chat(user, SPAN_WARNING("The securing bolts are not visible while maintenance protocols are disabled."))
+					return
+				user.visible_message(SPAN_NOTICE("\The [user] starts dismantling \the [src]..."), SPAN_NOTICE("You start disassembling \the [src]..."))
+				if(do_after(user, 30, TRUE, src))
+					user.visible_message(SPAN_NOTICE("\The [user] dismantles \the [src]."), SPAN_NOTICE("You disassemble \the [src]."))
+					dismantle()
+					return
 			else if(thing.iswelder())
 				if(!getBruteLoss())
 					return
